Refactor non-gradio stuff that's db/user related back to gpt_langchain.py
